+++ titus-ext/jooqflyway/gradle/dependency-locks/testAnnotationProcessor.lockfile
@@ -1,0 +1,3 @@
+# This is a Gradle generated file for dependency locking.
+# Manual edits can break the build and are not advised.
+# This file is expected to be part of source control.
