+++ titus-ext/jooqflyway/src/main/java/com/netflix/titus/ext/jooqflyway/JooqContext.java
@@ -1,4 +1,4 @@
-package com.netflix.titus.supplementary.jobactivity.store;
+package com.netflix.titus.ext.jooqflyway;
 
 import javax.sql.DataSource;
