+++ titus-ext/jooqflyway/src/main/java/com/netflix/titus/ext/jooqflyway/JooqUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.supplementary.jobactivity.store;
+package com.netflix.titus.ext.jooqflyway;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
