+++ titus-ext/jooqflyway/src/main/java/com/netflix/titus/ext/jooqflyway/RDSSSLSocketFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.supplementary.jobactivity.store;
+package com.netflix.titus.ext.jooqflyway;
 
 import org.postgresql.ssl.WrappedFactory;
