+++ titus-supplementary-component/job-activity-history/gradle/dependency-locks/testRuntimeClasspath.lockfile
@@ -33,7 +33,6 @@ com.github.jnr:jnr-ffi:2.0.7
 com.github.jnr:jnr-posix:3.0.27
 com.github.jnr:jnr-x86asm:1.0.2
 com.github.rholder:snowball-stemmer:1.3.0.581.1
-com.github.spotbugs:spotbugs-annotations:3.1.12
 com.github.spullara.cli-parser:cli-parser:1.1.1
 com.google.android:annotations:4.1.1.4
 com.google.api.grpc:proto-google-common-protos:1.12.0
@@ -79,7 +78,6 @@ com.netflix.runtime:health-core:1.1.4
 com.netflix.runtime:health-guice:1.1.4
 com.netflix.spectator:spectator-api:0.59.0
 com.ning:compress-lzf:0.8.4
-com.opentable.components:otj-pg-embedded:0.13.3
 com.squareup.okhttp3:logging-interceptor:3.14.9
 com.squareup.okhttp3:okhttp:3.14.9
 com.squareup.okio:okio:1.17.2
@@ -104,7 +102,6 @@ com.twitter:util-lint_2.11:6.32.0
 com.twitter:util-logging_2.11:6.32.0
 com.twitter:util-registry_2.11:6.32.0
 com.twitter:util-stats_2.11:6.32.0
-com.zaxxer:HikariCP:3.2.0
 commons-cli:commons-cli:1.3.1
 commons-codec:commons-codec:1.14
 commons-collections:commons-collections:3.2.1
@@ -173,7 +170,6 @@ jakarta.annotation:jakarta.annotation-api:1.3.5
 jakarta.xml.bind:jakarta.xml.bind-api:2.3.3
 janino:janino:2.5.10
 javax.activation:activation:1.1
-javax.activation:javax.activation-api:1.2.0
 javax.annotation:javax.annotation-api:1.3.2
 javax.el:javax.el-api:3.0.1-b06
 javax.inject:javax.inject:1
@@ -181,7 +177,6 @@ javax.mail:mailapi:1.4.3
 javax.servlet:javax.servlet-api:3.1.0
 javax.validation:validation-api:1.1.0.Final
 javax.ws.rs:jsr311-api:1.1.1
-javax.xml.bind:jaxb-api:2.3.1
 joda-time:joda-time:2.10.6
 junit:junit-dep:4.11
 junit:junit:4.13
@@ -232,7 +227,6 @@ org.eclipse.jetty:jetty-servlet:9.2.12.v20150709
 org.eclipse.jetty:jetty-util:9.2.12.v20150709
 org.eclipse.jetty:jetty-webapp:9.2.12.v20150709
 org.eclipse.jetty:jetty-xml:9.2.12.v20150709
-org.flywaydb:flyway-core:6.1.4
 org.fusesource:sigar:1.6.4
 org.glassfish:jakarta.el:3.0.3
 org.glassfish:javax.el:3.0.1-b12
@@ -244,9 +238,6 @@ org.hibernate:hibernate-validator:5.4.3.Final
 org.javassist:javassist:3.18.2-GA
 org.jboss.logging:jboss-logging:3.3.0.Final
 org.joda:joda-convert:2.2.1
-org.jooq:jooq-codegen:3.13.2
-org.jooq:jooq-meta:3.13.2
-org.jooq:jooq:3.13.2
 org.json:json:20140107
 org.junit.jupiter:junit-jupiter-api:5.6.3
 org.junit.jupiter:junit-jupiter-engine:5.6.3
@@ -271,13 +262,12 @@ org.ow2.asm:asm-commons:5.0.3
 org.ow2.asm:asm-tree:5.0.3
 org.ow2.asm:asm-util:5.0.3
 org.ow2.asm:asm:5.0.4
-org.postgresql:postgresql:42.2.8
 org.reactivestreams:reactive-streams:1.0.3
 org.reflections:reflections:0.9.10
 org.scala-lang.modules:scala-parser-combinators_2.11:1.0.4
 org.scala-lang:scala-library:2.11.7
 org.skyscreamer:jsonassert:1.5.0
-org.slf4j:jcl-over-slf4j:1.7.28
+org.slf4j:jcl-over-slf4j:1.7.7
 org.slf4j:jul-to-slf4j:1.7.30
 org.slf4j:slf4j-api:1.7.30
 org.slf4j:slf4j-log4j12:1.7.0
@@ -307,7 +297,6 @@ org.springframework:spring-webflux:5.2.11.RELEASE
 org.springframework:spring-webmvc:5.2.11.RELEASE
 org.synchronoss.cloud:nio-multipart-parser:1.1.0
 org.synchronoss.cloud:nio-stream-storage:1.1.3
-org.tukaani:xz:1.5
 org.webjars:swagger-ui:2.1.4
 org.xerial.snappy:snappy-java:1.1.8.4
 org.xmlunit:xmlunit-core:2.7.0
